10_cryptography_agent/pyproject.toml
[project]
name = "10-cryptography-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents[litellm]>=0.0.19",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
